COG-1107: NsUserService interface is extracted
